Feature/sp 4677 css cleanup (#34)
I moved all the styles in App.css to their respective module. I am not
sure though, about the "#root" style (file App.module.scss): apparently,
the map is not displayed if the root element is not set at 100% height,
and using a global style is the only way I found to define it in a
module.